+++ HISTORY.rst
@@ -25,6 +25,7 @@ latest (unreleased)
 
 * BIZ-878: Make cancellation of sales orders only synchronous
 * Changes Sales button on claim to include inactive SO
+* BIZ-736 add scenario to modify locations' partners
 
 **Bugfixes**